Export cache component keeping internal classes private
