+++ test/functional/http/get/cache.js
@@ -18,7 +18,7 @@ require('../../functional.js');
 const Promise = require('bluebird');
 const SwaggerEndpoint = require('../../../common/swagger_spec');
 const accountFixtures = require('../../../fixtures/accounts');
-const createCache = require('../../../../components');
+const { createCacheComponent } = require('../../../../components/cache');
 const Logger = require('../../../../logger');
 const apiHelpers = require('../../../common/helpers/api');
 const waitFor = require('../../../common/utils/wait_for');
@@ -37,7 +37,7 @@ describe('cached endpoints', () => {
 			errorLevel: __testContext.config.fileLogLevel,
 			filename: __testContext.config.logFileName,
 		});
-		cache = createCache(__testContext.config.redis, this.logger);
+		cache = createCacheComponent(__testContext.config.redis, this.logger);
 		await cache.bootstrap();
 		expect(cache).to.be.an('object');
 	});
